Working on animation (see full message)
The time definitions will need to be redone. Since we're allowing hour/day
increments/decrements we need to base the simulation time on a unit that
is a common divisor (e.g. a second) of these deltas instead of it being
based on the lunar synodic period.

Otherwise, if a user presses an increment or decrement hour button enough
times the displayed day may not be as expected (e.g. press increment hour
24 times and the day hasn't changed since T is just a tiny epsilon below
the transition value).